+++ exch/http/system_services.cpp
@@ -9,8 +9,8 @@
 using namespace gromox;
 
 decltype(system_services_judge_ip) system_services_judge_ip;
-BOOL (*system_services_judge_user)(const char*);
-BOOL (*system_services_add_user_into_temp_list)(const char *, int);
+bool (*system_services_judge_user)(const char *);
+void (*system_services_ban_user)(const char *, int);
 decltype(system_services_auth_login) system_services_auth_login;
 decltype(system_services_auth_meta) system_services_auth_meta;
 
@@ -26,7 +26,7 @@ int system_services_run()
 
 	E(system_services_judge_ip, "ip_filter_judge");
 	E(system_services_judge_user, "user_filter_judge");
-	E(system_services_add_user_into_temp_list, "user_filter_add");
+	E(system_services_ban_user, "user_filter_ban");
 	E(system_services_auth_login, "auth_login_gen");
 	E(system_services_auth_meta, "mysql_auth_meta");
 	return 0;
@@ -37,7 +37,7 @@ void system_services_stop()
 {
 	service_release("ip_filter_judge", "system");
 	service_release("user_filter_judge", "system");
-	service_release("user_filter_add", "system");
+	service_release("user_filter_ban", "system");
 	service_release("auth_login_gen", "system");
 	service_release("mysql_auth_meta", "system");
 }