+++ mda/smtp/smtp_aux.hpp
@@ -41,6 +41,4 @@ enum{
 
 };
 
-extern BOOL (*system_services_judge_user)(const char*);
-extern BOOL (*system_services_add_user_into_temp_list)(const char *, int);
 extern bool (*system_services_check_user)(const char *, const char *, char *, size_t);