+++ mra/imap/imap.hpp
@@ -203,8 +203,8 @@ extern void imrpc_build_env();
 extern void imrpc_free_env();
 
 extern bool (*system_services_judge_ip)(const char *host, std::string &reason);
-extern BOOL (*system_services_judge_user)(const char *);
-extern BOOL (*system_services_add_user_into_temp_list)(const char *, int);
+extern bool (*system_services_judge_user)(const char *);
+extern void (*system_services_ban_user)(const char *, int);
 extern authmgr_login_t system_services_auth_login;
 extern gromox::errno_t (*system_services_auth_meta)(const char *username, unsigned int wantpriv, sql_meta_result &out);
 extern int (*system_services_get_uid)(const char *, const char *, const std::string &, unsigned int *);