+++ mra/pop3/main.cpp
@@ -45,7 +45,7 @@ using namespace gromox;
 #define E(s) decltype(system_services_ ## s) system_services_ ## s;
 E(judge_ip)
 E(judge_user)
-E(add_user_into_temp_list)
+E(ban_user)
 E(auth_login)
 E(auth_meta)
 E(list_mail)
@@ -163,7 +163,7 @@ static int system_services_run()
 
 	E(system_services_judge_ip, "ip_filter_judge");
 	E(system_services_judge_user, "user_filter_judge");
-	E(system_services_add_user_into_temp_list, "user_filter_add");
+	E(system_services_ban_user, "user_filter_ban");
 	E(system_services_auth_login, "auth_login_gen");
 	E(system_services_auth_meta, "mysql_auth_meta");
 	E(system_services_list_mail, "list_mail");
@@ -177,7 +177,7 @@ static void system_services_stop()
 {
 	service_release("ip_filter_judge", "system");
 	service_release("user_filter_judge", "system");
-	service_release("user_filter_add", "system");
+	service_release("user_filter_ban", "system");
 	service_release("mysql_auth_meta", "system");
 	service_release("auth_login_gen", "system");
 	service_release("list_mail", "system");