+++ mra/pop3/pop3.hpp
@@ -100,8 +100,8 @@ extern void xrpc_build_env();
 extern void xrpc_free_env();
 
 extern bool (*system_services_judge_ip)(const char *host, std::string &reason);
-extern BOOL (*system_services_judge_user)(const char *);
-extern BOOL (*system_services_add_user_into_temp_list)(const char *, int);
+extern bool (*system_services_judge_user)(const char *);
+extern void (*system_services_ban_user)(const char *, int);
 extern authmgr_login_t system_services_auth_login;
 extern gromox::errno_t (*system_services_auth_meta)(const char *username, unsigned int wantpriv, sql_meta_result &out);
 extern int (*system_services_list_mail)(const char *, const char *, std::vector<gromox::MSG_UNIT> &, int *num, uint64_t *size);